removed side requirements, added parent
Removed all side requirements for fissions and removed pseudo fissions
(de facto). Will make branch paraphyly-alt which includes pseudo
fissions in the process - relevant for model selection. This way the
tumour is not guaranteed to have 4 demes on the left and 4 demes on the
right side. The labels don't really matter, but are useful. Also added
tracking of deme parent, can easily construct trees now.